requirements.txt
rasa_core==0.13.2
rasa_nlu==0.14.3
Django==2.2
rasa_core_sdk==0.12.1
python_dateutil==2.8.0
duckling==1.8.0
djangorestframework==3.9.2
